+++ lib/rmodbus/rtu_server.rb
@@ -11,8 +11,7 @@
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
-require 'serialport'
-
+#
 module ModBus
   # RTU server implementation
   # @example